+++ battle/attack/AttackSimple.java
@@ -181,7 +181,7 @@ public class AttackSimple extends AttackAb {
 		if (proc.MINIVOLC.prob > 0 && (!capt.isEmpty() || proc.MINIVOLC.hitless)) {
 			int dire = model.getDire();
 			Proc.MINIVOLC volc = proc.MINIVOLC;
-			int addp = volc.dis_0 == volc.dis_1 ? volc.dis_0 : volc.dis_0 + (int) (model.b.r.nextDouble() * (volc.dis_1 - volc.dis_0));
+			int addp = volc.dis_0 == volc.dis_1 ? volc.dis_0 : volc.dis_0 + (int) (model.b.r.nextFloat() * (volc.dis_1 - volc.dis_0));
 			float p0 = model.getPos() + dire * addp;
 			float sta = p0 + (dire == 1 ? W_VOLC_PIERCE : W_VOLC_INNER);
 			float end = p0 - (dire == 1 ? W_VOLC_INNER : W_VOLC_PIERCE);
